Don't add radicals or lone pairs to surface sites.
This may hide bugs, or it may be exactly the right
thing to do. I'm not entirely sure.
But it currently seems like a good idea.

The presumption is:
surface sites of metals should not have unpaired
or pairs of electrons - they just have a big sea
of delocalized electrons at their disposal.

See https://github.com/ReactionMechanismGenerator/RMG-Py/issues/1820