Don't UNMARK duplicate reactions so aggressively.
There's a model from Klippenstein/Glarborg, with three
duplicate reactions, two PDep and one not.
The previous implementation would notice that because 
one is PDep and one is not, it would REMOVE the DUPLICATE
flag, even though there's a THIRD reaction that means 
the DUPLICATE flag should remain.

Now we just log a warning and leave the flag alone.